pyproject.toml
[tool.poetry]
name = "ml-data-pipeline"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.13.0"
kfp = "^2.7.0"
google-cloud-aiplatform = "^1.48.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
